contrib: movpatch, gradmask, forecast_weights with flat end, data_loadings with new options
